refactor: refactor backend
