+++ server/src/api/entries.ts
@@ -1,7 +1,6 @@
 import express from "express";
 
-import { updateEntryStatus } from "../controllers/updateEntry.js";
-import { verifyToken } from "../controllers/verifyToken.js";
+import { updateEntryStatus, verifyToken } from "../controllers/index.js";
 
 const router = express.Router();
