+++ server/src/api/login.ts
@@ -1,6 +1,6 @@
 import express from "express";
 
-import { login } from "../controllers/login.js";
+import { login } from "../controllers/index.js";
 
 const router = express.Router();
